refactor(ports): extract abstract bases into hyperion/ports
Closes #126
