+++ hyperion/catalog/__init__.py
@@ -1,5 +1,6 @@
+from hyperion.ports.schema_registry import SchemaStore
+
 from .catalog import AssetNotFoundError, Catalog
-from .schema import SchemaStore
 
 __all__ = [
     "AssetNotFoundError",
