+++ hyperion/infrastructure/cache.py
@@ -1,279 +1,72 @@
-"""A serverless cache for our shenanigans."""
+"""Concrete cache adapters.
 
-import hashlib
+.. deprecated::
+    The abstract :class:`Cache`, :class:`CacheStats` and :class:`CachingError`
+    moved to :mod:`hyperion.ports.cache`. Import them from there. This module
+    keeps them importable (with a :class:`DeprecationWarning`) for the whole
+    ``hyperion-sdk`` 1.x line and still hosts the concrete adapters until S6.
+"""
+
 import tempfile
 import time
-from abc import ABC, abstractmethod
 from base64 import b64decode, b64encode
 from collections.abc import Iterator
 from contextlib import contextmanager
-from dataclasses import dataclass, replace
-from io import BytesIO, StringIO
 from pathlib import Path
-from typing import IO, Any, ClassVar, Literal, cast, overload
+from typing import IO, TYPE_CHECKING, Any, Literal, cast, overload
 
 import boto3
 import cachetools
-import snappy
 
+from hyperion._compat import moved_attr
 from hyperion.catalog import AssetNotFoundError, Catalog
-from hyperion.config import storage_config
 from hyperion.dateutils import utcnow
 from hyperion.entities.catalog import PersistentStoreAsset
 from hyperion.log import get_logger
+from hyperion.ports.cache import DEFAULT_TTL_SECONDS, CacheKeyOpenMode
+from hyperion.ports.cache import Cache as _Cache
+from hyperion.ports.cache import CacheStats as _CacheStats
+from hyperion.ports.cache import CachingError as _CachingError
 
-DEFAULT_TTL_SECONDS = 60
+if TYPE_CHECKING:
+    from hyperion.ports.cache import Cache, CacheStats, CachingError
+
 DYNAMODB_MAX_LENGTH = 65535
 DEFAULT_LOCAL_FILE_CACHE_MAX_SIZE = 256 * (1024**2)
 
-CacheKeyOpenMode = Literal["str", "bytes"]
-
 logger = get_logger("cache")
 
+_MOVED: dict[str, tuple[object, str]] = {
+    "Cache": (_Cache, "hyperion.ports.cache"),
+    "CacheStats": (_CacheStats, "hyperion.ports.cache"),
+    "CachingError": (_CachingError, "hyperion.ports.cache"),
+}
 
-@dataclass
-class CacheStats:
-    hits: int
-    misses: int
-    gets: int
-    deletes: int
-    clears: int
-    sets: int
-
-    @classmethod
-    def empty(cls) -> "CacheStats":
-        return cls(hits=0, misses=0, gets=0, deletes=0, clears=0, sets=0)
-
-    def __sub__(self, other: Any) -> dict[str, int]:
-        if not isinstance(other, CacheStats):
-            return NotImplemented
-        difference: dict[str, int] = {}
-        for field in CacheStats.__dataclass_fields__:
-            ours = getattr(self, field)
-            theirs = getattr(other, field)
-            if not isinstance(ours, int):
-                raise TypeError(f"Field {field!r} is not an integer on {self!r}.")
-            if not isinstance(theirs, int):
-                raise TypeError(f"Field {field!r} is not an integer on {other!r}.")
-            if diff := ours - theirs:
-                difference[field] = diff
-        return difference
+__all__ = [
+    "DEFAULT_LOCAL_FILE_CACHE_MAX_SIZE",
+    "DEFAULT_TTL_SECONDS",
+    "DYNAMODB_MAX_LENGTH",
+    "Cache",
+    "CacheKeyOpenMode",
+    "CacheStats",
+    "CachingError",
+    "DynamoDBCache",
+    "InMemoryCache",
+    "LocalFileCache",
+    "PersistentCache",
+]
 
 
-class CachingError(Exception):
-    pass
-
-
-class Cache(ABC):
-    """A serverless cache for our shenanigans."""
-
-    _instances: ClassVar[dict[tuple[str, bool], "Cache"]] = {}
-
-    @classmethod
-    def from_config(cls) -> "Cache":
-        """
-        Creates a cache from the configuration.
-
-        This function emulates a singleton pattern, so it will return the same instance for the same configuration.
-
-        Returns:
-            Cache: A cache instance.
-        """
-        instance_key = (storage_config.cache_key_prefix, True)
-        if instance_key not in Cache._instances:
-            if storage_config.cache_dynamodb_table:
-                logger.info("Using DynamoDB Cache.")
-                cls._instances[instance_key] = DynamoDBCache(
-                    prefix=storage_config.cache_key_prefix,
-                    default_ttl=storage_config.cache_dynamodb_default_ttl,
-                    table_name=storage_config.cache_dynamodb_table,
-                )
-            elif storage_config.cache_local_path:
-                logger.info("Using LocalFileCache.", path=storage_config.cache_local_path)
-                cls._instances[instance_key] = LocalFileCache(
-                    prefix=storage_config.cache_key_prefix, root_path=Path(storage_config.cache_local_path)
-                )
-            else:
-                logger.info("Using InMemory Cache.")
-                cls._instances[instance_key] = InMemoryCache(
-                    prefix=storage_config.cache_key_prefix, default_ttl=storage_config.cache_dynamodb_default_ttl
-                )
-        return cls._instances[instance_key]
-
-    def __init__(self, prefix: str, hash_keys: bool = True, default_ttl: int = DEFAULT_TTL_SECONDS):
-        """Initializes the cache with the given prefix and default TTL.
-
-        Args:
-            prefix (str): The prefix for the cache keys.
-            hash_keys (bool): Whether to hash the keys.
-            default_ttl (int): The default TTL for the cache.
-        """
-        self.prefix = prefix
-        self.hash_keys = hash_keys
-        self.default_ttl = default_ttl
-        self._stats = CacheStats.empty()
-
-    def reset_stats(self) -> None:
-        self._stats = CacheStats.empty()
-
-    @property
-    def stats(self) -> CacheStats:
-        return replace(self._stats)
-
-    def _key(self, key: str) -> str:
-        """Generates a cache key from a given key."""
-        if self.hash_keys:
-            key = hashlib.sha256(key.encode(encoding="utf-8")).hexdigest()
-        prefix = f"{self.prefix}:" if self.prefix else ""
-        return f"{prefix}{key}"
-
-    def _compress(self, value: str) -> bytes:
-        """Compresses a value using snappy compression."""
-        return self._compress_bytes(value.encode(encoding="utf-8"))
-
-    def _compress_bytes(self, value: bytes) -> bytes:
-        """Compress a bytes value using snappy compression."""
-        compressed_value = snappy.compress(value)
-        logger.debug(
-            "Compressed value using snappy compression.",
-            original_length=len(value),
-            compressed_length=len(compressed_value),
-            ratio=len(compressed_value) / len(value),
+def __getattr__(name: str) -> object:
+    if name in _MOVED:
+        value, new_module = _MOVED[name]
+        return moved_attr(
+            name=name, value=value, old_module="hyperion.infrastructure.cache", new_module=new_module
         )
-        return cast(bytes, compressed_value)
-
-    def _decompress(self, value: bytes) -> str:
-        """Decompresses a value using snappy decompression."""
-        return self._decompress_bytes(value).decode(encoding="utf-8")
-
-    def _decompress_bytes(self, value: bytes) -> bytes:
-        """Decompresses a bytes value using snappy decompression."""
-        return cast(bytes, snappy.decompress(value))
-
-    @overload
-    @contextmanager
-    def _open(self, key: str, mode: Literal["str"]) -> Iterator[IO[str]]:
-        pass
-
-    @overload
-    @contextmanager
-    def _open(self, key: str, mode: Literal["bytes"]) -> Iterator[IO[bytes]]:
-        pass
-
-    @contextmanager
-    def _open(self, key: str, mode: CacheKeyOpenMode = "str") -> Iterator[IO[bytes] | IO[str]]:
-        """Opens a file-like object for reading and writing to the cache.
-        This is useful for streaming data to and from the cache.
-        In the base class, this has no performance benefits, because the implementation is in memory (using
-        StringIO or BytesIO), but in the LocalFileCache, this is useful for streaming data to and from the file system.
-
-        Args:
-            key (str): The key for the cache.
-            mode (str): The mode for opening the file-like object. Can be "str" or "bytes".
-
-        Yields:
-            IO[bytes] | IO[str]: A file-like object for reading and writing to the cache.
-        """
-        if mode == "str":
-            with StringIO(self._get(key) or "") as file:
-                yield file
-                self._set(key, file.getvalue())
-        elif mode == "bytes":
-            with BytesIO(self._get_bytes(key) or b"") as file:
-                yield file
-                self._set_bytes(key, file.getvalue())
-        else:
-            raise ValueError(f"Unsupported open mode {mode!r} - 'str' or 'bytes' are supported.")
-
-    @overload
-    @contextmanager
-    def open(self, key: str, mode: Literal["str"]) -> Iterator[IO[str]]:
-        pass
-
-    @overload
-    @contextmanager
-    def open(self, key: str, mode: Literal["bytes"]) -> Iterator[IO[bytes]]:
-        pass
-
-    @contextmanager
-    def open(self, key: str, mode: CacheKeyOpenMode = "str") -> Iterator[IO[bytes] | IO[str]]:
-        self._stats.gets += 1
-        with self._open(key, mode) as file:
-            yield file
-        self._stats.sets += 1
-
-    @abstractmethod
-    def _get(self, key: str) -> str | None:
-        pass
-
-    def get(self, key: str) -> str | None:
-        """Gets a value from the cache."""
-        self._stats.gets += 1
-        if (cached := self._get(key)) is not None:
-            self._stats.hits += 1
-        else:
-            self._stats.misses += 1
-        return cached
-
-    @abstractmethod
-    def _set(self, key: str, value: str) -> None:
-        pass
-
-    def set(self, key: str, value: str) -> None:
-        """Sets a value in the cache."""
-        self._stats.sets += 1
-        self._set(key, value)
-
-    @abstractmethod
-    def _get_bytes(self, key: str) -> bytes | None:
-        pass
-
-    def get_bytes(self, key: str) -> bytes | None:
-        """Gets a bytes value from the cache."""
-        self._stats.gets += 1
-        if (cached := self._get_bytes(key)) is not None:
-            self._stats.hits += 1
-        else:
-            self._stats.misses += 1
-        return cached
-
-    @abstractmethod
-    def _set_bytes(self, key: str, value: bytes) -> None:
-        pass
-
-    def set_bytes(self, key: str, value: bytes) -> None:
-        """Sets a bytes value in the cache."""
-        self._stats.sets += 1
-        self._set_bytes(key, value)
-
-    @abstractmethod
-    def _delete(self, key: str) -> None:
-        pass
-
-    def delete(self, key: str) -> None:
-        """Deletes a value from the cache."""
-        self._stats.deletes += 1
-        self._delete(key)
-
-    @abstractmethod
-    def _clear(self) -> None:
-        pass
-
-    def clear(self) -> None:
-        """Clears the cache."""
-        self._stats.clears += 1
-        self._clear()
-
-    @abstractmethod
-    def _hit(self, key: str) -> bool:
-        pass
-
-    def hit(self, key: str) -> bool:
-        """Checks if a key exists in the cache."""
-        return self._hit(key)
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
 
 
-class InMemoryCache(Cache):
+class InMemoryCache(_Cache):
     """An in-memory cache for our shenanigans."""
 
     MAX_KEYS = 1000
@@ -301,7 +94,7 @@ class InMemoryCache(Cache):
         try:
             return b64decode(cached.encode("ascii"))
         except Exception as error:
-            raise CachingError(f"Failed to decode cached key {key!r}, make sure it was stored as bytes.") from error
+            raise _CachingError(f"Failed to decode cached key {key!r}, make sure it was stored as bytes.") from error
 
     def _delete(self, key: str) -> None:
         self.cache.pop(self._key(key), None)
@@ -313,7 +106,7 @@ class InMemoryCache(Cache):
         return self._key(key) in self.cache
 
 
-class LocalFileCache(Cache):
+class LocalFileCache(_Cache):
     """A local file cache for our shenanigans."""
 
     def __init__(
@@ -472,7 +265,7 @@ class LocalFileCache(Cache):
             file.unlink()
 
 
-class PersistentCache(Cache):
+class PersistentCache(_Cache):
     """Uses a persistent store asset to store the cached data.
 
     Persistent store asset is a key-value store that is stored in the catalog.
@@ -553,7 +346,7 @@ class PersistentCache(Cache):
         try:
             return b64decode(cached.encode("ascii"))
         except Exception as error:
-            raise CachingError(f"Failed to decode cached key {key!r}, make sure it was stored as bytes.") from error
+            raise _CachingError(f"Failed to decode cached key {key!r}, make sure it was stored as bytes.") from error
 
     def _delete(self, key: str) -> None:
         cache_key = self._key(key)
@@ -567,7 +360,7 @@ class PersistentCache(Cache):
         return self._key(key) in self.data
 
 
-class DynamoDBCache(Cache):
+class DynamoDBCache(_Cache):
     """A DynamoDB cache for our shenanigans."""
 
     TTL_ATTRIBUTE_NAME = "time_to_live"
@@ -601,7 +394,7 @@ class DynamoDBCache(Cache):
                 cache_key=cache_key,
                 length=len(compressed_value),
             )
-            raise CachingError(f"Value is too long to store in DynamoDB: {len(compressed_value)}")
+            raise _CachingError(f"Value is too long to store in DynamoDB: {len(compressed_value)}")
         self.table.put_item(
             Item={"key": cache_key, "value": compressed_value, self.TTL_ATTRIBUTE_NAME: expiration_time}
         )
