+++ hyperion/infrastructure/keyval.py
@@ -1,200 +1,57 @@
-"""Key-Value stores.
+"""Concrete key-value store adapters.
 
-If you need a store with a TTL, look at Cache instead.
+If you need a store with a TTL, look at :class:`hyperion.ports.cache.Cache`
+instead.
+
+.. deprecated::
+    The abstract :class:`KeyValueStore` and the :data:`CompressionType` alias
+    moved to :mod:`hyperion.ports.keyval`. Import them from there. This module
+    keeps them importable (with a :class:`DeprecationWarning`) for the whole
+    ``hyperion-sdk`` 1.x line and still hosts the concrete adapters until S6.
 """
 
-import gzip
-from abc import ABC, abstractmethod
-from collections.abc import Iterable, Iterator
-from fnmatch import fnmatch
-from typing import TYPE_CHECKING, Any, Literal, TypeGuard, cast
+from collections.abc import Iterable
+from typing import TYPE_CHECKING, Any, TypeGuard, cast
 
 import boto3
-import snappy
+
+from hyperion._compat import moved_attr
+from hyperion.ports.keyval import CompressionType as _CompressionType
+from hyperion.ports.keyval import KeyValueStore as _KeyValueStore
 
 if TYPE_CHECKING:
     from mypy_boto3_dynamodb.type_defs import ScanInputTableScanTypeDef
 
-CompressionType = Literal["snappy", "gzip"]
+    from hyperion.ports.keyval import CompressionType, KeyValueStore
+
+_MOVED: dict[str, tuple[object, str]] = {
+    "KeyValueStore": (_KeyValueStore, "hyperion.ports.keyval"),
+    "CompressionType": (_CompressionType, "hyperion.ports.keyval"),
+}
+
+__all__ = [
+    "CompressionType",
+    "DynamoDBStore",
+    "InMemoryStore",
+    "KeyValueStore",
+    "is_valid_compression_type",
+]
 
 
-def is_valid_compression_type(compression_type: str) -> TypeGuard[CompressionType]:
+def __getattr__(name: str) -> object:
+    if name in _MOVED:
+        value, new_module = _MOVED[name]
+        return moved_attr(
+            name=name, value=value, old_module="hyperion.infrastructure.keyval", new_module=new_module
+        )
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+
+
+def is_valid_compression_type(compression_type: str) -> TypeGuard[_CompressionType]:
     return compression_type in ("snappy", "gzip")
 
 
-class KeyValueStore(ABC, Iterable[str]):
-    """A key-value store."""
-
-    def __init__(self, prefix: str | None = None, compression: CompressionType | None = None) -> None:
-        """Initialize the store.
-
-        Args:
-            prefix: A prefix to add to all keys.
-            compression: The compression algorithm to use.
-        """
-        self.prefix = prefix
-        self.compression = compression
-
-    def _key(self, key: str) -> str:
-        """Return the key with the prefix added."""
-        prefix = f"{self.prefix}:" if self.prefix else ""
-        return f"{prefix}{key}"
-
-    def _compress(self, value: str) -> bytes:
-        """Compress a value using configured compression.
-
-        If no compression is selected, the string is encoded as utf-8 into bytes.
-
-        Args:
-            value: The value to compress.
-
-        Returns:
-            bytes: The compressed value.
-
-        Raises:
-            ValueError: If an unsupported compression is selected.
-        """
-        value_bytes = value.encode("utf-8")
-        match self.compression:
-            case None:
-                return value_bytes
-            case "snappy":
-                return cast(bytes, snappy.compress(value_bytes))
-            case "gzip":
-                return gzip.compress(value_bytes)
-            case _:
-                raise ValueError(f"Unsupported compression {self.compression!r}.")
-
-    def _decompress(self, value: bytes) -> str:
-        """Decompress a value using configured compression.
-
-        If no compression is selected, the bytes are decoded as utf-8 into a string.
-
-        Args:
-            value: The value to decompress.
-
-        Returns:
-            str: The decompressed value.
-
-        Raises:
-            ValueError: If an unsupported compression is selected.
-            TypeError: If the decompression returns an unexpected type.
-        """
-        match self.compression:
-            case None:
-                return value.decode("utf-8")
-            case "snappy":
-                value_decompressed = snappy.decompress(value)
-                if isinstance(value_decompressed, str):
-                    return value_decompressed
-                elif isinstance(value_decompressed, bytes):
-                    return value_decompressed.decode("utf-8")
-                raise TypeError(
-                    "Unexpected value returned from snappy decompression, "
-                    f"expected str | bytes, got {type(value_decompressed)!r}."
-                )
-            case "gzip":
-                return gzip.decompress(value).decode("utf-8")
-            case _:
-                raise ValueError(f"Unsupported compression {self.compression!r}.")
-
-    @abstractmethod
-    def _get_raw(self, hashed_key: str) -> bytes | None:
-        """Get a raw uncompressed value from the store.
-
-        The key passed here must already be hashed and prefixed.
-
-        Args:
-            hashed_key: The key to get.
-
-        Returns:
-            bytes | None: The raw value, or None if not found.
-        """
-
-    def get(self, key: str) -> str | None:
-        """Gets a value from the store."""
-        if value := self._get_raw(self._key(key)):
-            return self._decompress(value)
-        return None
-
-    @abstractmethod
-    def _set_raw(self, hashed_key: str, compresed_value: bytes) -> None:
-        """Set a raw, compressed bytes value in the store.
-
-        The key passed here must already by hashed and prefixed.
-
-        Args:
-            hashed_key: The key to set.
-            compresed_value: The compressed value to set.
-        """
-
-    def set(self, key: str, value: str) -> None:
-        """Sets a value in the store."""
-        return self._set_raw(self._key(key), self._compress(value))
-
-    @abstractmethod
-    def _delete_raw(self, hashed_key: str) -> None:
-        """Delete a value from the store.
-
-        The key passed here must already by hashed and prefixed.
-
-        Args:
-            hashed_key: The key to delete.
-        """
-
-    def __iter__(self) -> Iterator[str]:
-        """Iterate all keys available in the store."""
-        return iter(self._iter_all_keys())
-
-    @abstractmethod
-    def _iter_all_keys(self) -> Iterable[str]:
-        """Iterate all keys available in the store."""
-
-    def keys(self, match: str = "*") -> Iterable[str]:
-        """Iterate all keys available in the store that match the given expression.
-
-        Nothing fancy is supported, UNIX-style fnmatch is performed.
-
-        Args:
-            match: The expression to match.
-
-        Returns:
-            Iterable[str]: The keys that match the expression
-        """
-        for key in self._iter_all_keys():
-            if fnmatch(key, match):
-                yield key
-
-    def delete(self, key: str) -> None:
-        """Delete a value from the store.
-
-        Args:
-            key: The key to delete.
-        """
-        return self._delete_raw(self._key(key))
-
-    def exists(self, key: str) -> bool:
-        """Returns whether a key exists in the store.
-
-        Args:
-            key: The key to check.
-        """
-        return key in self.keys()
-
-    def __contains__(self, key: str) -> bool:
-        return self.exists(key)
-
-    def __getitem__(self, key: str) -> str | None:
-        return self.get(key)
-
-    def __setitem__(self, key: str, value: str) -> None:
-        return self.set(key, value)
-
-    def __delitem__(self, key: str) -> None:
-        return self.delete(key)
-
-
-class DynamoDBStore(KeyValueStore):
+class DynamoDBStore(_KeyValueStore):
     """A key-value store using DynamoDB.
 
     The table must have a key attribute and a value attribute.
@@ -203,7 +60,7 @@ class DynamoDBStore(KeyValueStore):
     def __init__(
         self,
         prefix: str | None = None,
-        compression: CompressionType | None = None,
+        compression: _CompressionType | None = None,
         table_name: str | None = None,
         key_attribute: str = "key",
         value_attribute: str = "value",
@@ -247,13 +104,13 @@ class DynamoDBStore(KeyValueStore):
             scan_kwargs["ExclusiveStartKey"] = response["LastEvaluatedKey"]
 
 
-class InMemoryStore(KeyValueStore):
+class InMemoryStore(_KeyValueStore):
     """An in-memory key-value store.
 
     This store is not persistent and is not shared between instances.
     """
 
-    def __init__(self, prefix: str | None = None, compression: CompressionType | None = None):
+    def __init__(self, prefix: str | None = None, compression: _CompressionType | None = None):
         super().__init__(prefix, compression)
         self.store: dict[str, bytes] = {}
